+++ R/exportedMethods.R
@@ -222,14 +222,15 @@ setMethod(
 #' b <- "NULL"
 #' .prepareOutput(b) # converts to NULL
 #'
-#' library(raster)
-#' r <- raster(extent(0,10,0,10), vals = 1:100)
+#' if (requireNamespace("raster")) {
+#'   r <- raster(extent(0,10,0,10), vals = 1:100)
 #'
-#' # write to disk manually -- will be in tempdir()
-#' r <- writeRaster(r, file = tempfile())
+#'   # write to disk manually -- will be in tempdir()
+#'   r <- writeRaster(r, file = tempfile())
 #'
-#' # copy it to the cache repository
-#' r <- .prepareOutput(r, tempdir())
+#'   # copy it to the cache repository
+#'   r <- .prepareOutput(r, tempdir())
+#' }
 setGeneric(".prepareOutput", function(object, cachePath, ...) {
   standardGeneric(".prepareOutput")
 })